Improve performances when displaying revision info
The pane related to revision displays many informations. To do this, many git commands are launched. Such strategy is CPU/time consuming.

As the user only read a single tab at a time, we only update the tab displayed. When the user selects another tab, we update this tab if it is marked "dirty", ie a new revision was selected.